[SYSML-375] [Optimizer] Address Optimizer Inconsistencies Due to Casting of Doubles To Ints
This addresses a few inconsistencies with our optimizer in which accidental truncation of double values to integer values via casting could lead to rewrite rules being mistakenly applied.  As an example, a rewrite rule targeting the expression 'sum(v^2)' could be mistakenly applied to the expression 'sum(v^2.3)'.

Closes #2.

+++ src/main/java/com/ibm/bi/dml/hops/TernaryOp.java
@@ -930,7 +930,7 @@ public class TernaryOp extends Hop
 						DataGenOp dgop = (DataGenOp) input1;
 						if( dgop.getOp() == DataGenMethod.SEQ ){
 							Hop incr = dgop.getInput().get(dgop.getParamIndex(Statement.SEQ_INCR));
-							ret = (incr instanceof LiteralOp && HopRewriteUtils.getIntValue((LiteralOp)incr)==1)
+							ret = (incr instanceof LiteralOp && HopRewriteUtils.getDoubleValue((LiteralOp)incr)==1)
 								  || dgop.getIncrementValue()==1.0; //set by recompiler
 						}
 					}
@@ -940,7 +940,7 @@ public class TernaryOp extends Hop
 						DataGenOp dgop = (DataGenOp) input2;
 						if( dgop.getOp() == DataGenMethod.SEQ ){
 							Hop incr = dgop.getInput().get(dgop.getParamIndex(Statement.SEQ_INCR));
-							ret |= (incr instanceof LiteralOp && HopRewriteUtils.getIntValue((LiteralOp)incr)==1)
+							ret |= (incr instanceof LiteralOp && HopRewriteUtils.getDoubleValue((LiteralOp)incr)==1)
 								   || dgop.getIncrementValue()==1.0; //set by recompiler;
 						}
 					}
